+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
+- added error messages for invalid upgrade and multiplier config entries
 - fixed grinder and infuser descriptions in German translation
 
 ## [3.1.0] - 2022-07-21